bst: re-format the code to indent 2 spaces
I'm sorry that I formated code to 4 spaces indent in commit ac81be37798.
So there are many alignment issues in the code which are very annoying.
Instead of fix the code one by one, I prefer to go back to indent 2
spaces. I have had a look at authordate1.bst plainDemo.bst, tex-live.bst
and IEEEtran.bst shipped with TexLive2012 and they are all indent with 2
spaces. So, although 2 spaces are too small IMHO, I set the indent to 2
spaces.

I hope this is the last time I deal with coding styles for the bst
file...